Resolve circular deps
